Played; reset to use AddPersonForm; Signup redirects incorrectly; can't add food

+++ src/main/java/org/launchcode/foodday/controllers/FooddayController.java
@@ -29,8 +29,13 @@ public class FooddayController {
 
 
     @RequestMapping(value = "")
-    public String index(Model model) {
+    public String index(Model model, @CookieValue(value="user", defaultValue = "none") String name) {
 
+        if(name.equals("none")) {
+            return "redirect:/user";
+        }
+
+        //User u = userDao.findByName(name).get(0);
         model.addAttribute("title", "Food Days");
         model.addAttribute("dates", fooddayDao.findAll());
 
@@ -50,7 +55,11 @@ public class FooddayController {
 
     @RequestMapping(value = "add", method = RequestMethod.POST)
     public String processAddDateForm(@ModelAttribute @Valid Foodday date, Errors errors,
-                                     Model model) {
+                                     Model model, @CookieValue(value="user", defaultValue = "none") String name) {
+
+        if(name.equals("none")) {
+            return "redirect:/user";
+        }
 
         if (errors.hasErrors()) {
             model.addAttribute("title", "Add Date");
@@ -73,7 +82,12 @@ public class FooddayController {
 
 
     @RequestMapping(value = "remove", method = RequestMethod.POST)
-    public String processRemoveDateForm(@RequestParam int[] dateIds) {
+    public String processRemoveDateForm(@RequestParam int[] dateIds,
+                                        @CookieValue(value="user", defaultValue = "none") String name) {
+
+        if(name.equals("none")) {
+            return "redirect:/user";
+        }
 
         for (int dateId : dateIds) {
             fooddayDao.delete(dateId);
@@ -84,7 +98,12 @@ public class FooddayController {
 
 
     @RequestMapping(value = "view/{dateId}", method = RequestMethod.GET)
-    public String viewDate(Model model, @PathVariable int dateId) {
+    public String viewDate(Model model, @PathVariable int dateId,
+                           @CookieValue(value="user", defaultValue = "none") String name) {
+
+        if(name.equals("none")) {
+            return "redirect:/user";
+        }
 
         Foodday date = fooddayDao.findOne(dateId);
         model.addAttribute("date", date);
@@ -95,35 +114,43 @@ public class FooddayController {
     }
 
 
-    @RequestMapping(value = "add-person/{dateId}", method = RequestMethod.GET)
+    @RequestMapping(value = "add-food/{dateId}", method = RequestMethod.GET)
     public String displayAddUserForm(Model model, @PathVariable int dateId,
                                      @ModelAttribute String food) {
 
         Foodday date = fooddayDao.findOne(dateId);
-        AddUserForm form = new AddUserForm(date, userDao.findAll(), food);
+        //List<User> users = date.getUsers();
+        AddUserForm form = new AddUserForm(userDao.findAll(), food, date);
 
         model.addAttribute("title", "Add Person to " + date.getDate());
-        model.addAttribute("date", date);
+        //model.addAttribute("date", date);
+        //model.addAttribute("users", users);
         model.addAttribute("form", form);
 
         return "date/add-person";
     }
 
 
-    @RequestMapping(value = "add-person", method = RequestMethod.POST)
+    @RequestMapping(value = "add-food", method = RequestMethod.POST)
     public String processAddUserForm(@ModelAttribute @Valid AddUserForm form, Errors errors,
-                                     @ModelAttribute String food, Model model) {
+                                     @ModelAttribute String food, Model model,
+                                     @CookieValue(value="user", defaultValue = "none") String name) {
+
+        if(name.equals("none")) {
+            return "redirect:/user";
+        }
 
         if (errors.hasErrors()) {
             model.addAttribute("title", "Add Person");
+            //model.addAttribute("date", date);
             //model.addAttribute("users", userDao.findAll());
+            //model.addAttribute("users", date.getUsers());
             model.addAttribute("form", form);
 
             return "date/add-person";
         }
 
         //Foodday date = fooddayDao.findOne(id);
-        //List<User> users = date.getUsers();
         User aUser = userDao.findOne(form.getUserId());
         Foodday aDate = fooddayDao.findOne(form.getDateId());
         aDate.addPerson(aUser, food);
@@ -133,7 +160,7 @@ public class FooddayController {
     }
 
 
-    @RequestMapping(value = "remove-person/{dateId}", method = RequestMethod.GET)
+    @RequestMapping(value = "remove-food/{dateId}", method = RequestMethod.GET)
     public String displayRemovePersonForm(Model model) {
         model.addAttribute("title", "Remove Person");
         model.addAttribute("persons", userDao.findAll());
@@ -142,9 +169,14 @@ public class FooddayController {
     }
 
 
-    @RequestMapping(value = "remove-person", method = RequestMethod.POST)
+    @RequestMapping(value = "remove-food", method = RequestMethod.POST)
     public String processRemovePersonForm(@RequestParam int[] userIds,
-                                          @ModelAttribute Foodday date) {
+                                          @ModelAttribute Foodday date,
+                                          @CookieValue(value="user", defaultValue = "none") String name) {
+        if(name.equals("none")) {
+            return "redirect:/user";
+        }
+
         for (int userId : userIds) {
 
             userDao.delete(userId);
@@ -154,13 +186,13 @@ public class FooddayController {
     }
 
 
-    @RequestMapping(value = "edit/{personId}", method = RequestMethod.GET)
+    @RequestMapping(value = "edit-food/{personId}", method = RequestMethod.GET)
     public String displayEditForm(Model model, @PathVariable int dateId,
                                   @RequestParam int userId, @ModelAttribute String food) {
 
         User aUser = userDao.findOne(userId);
         Foodday date = fooddayDao.findOne(dateId);
-        AddUserForm form = new AddUserForm(date, userDao.findAll(), food);
+        //AddUserForm form = new AddUserForm(userDao.findAll(), food, date);
 
         if (aUser == null) {
             return "redirect:view/";
@@ -172,10 +204,15 @@ public class FooddayController {
     }
 
 
-    @RequestMapping(value = "edit", method = RequestMethod.POST)
+    @RequestMapping(value = "edit-food", method = RequestMethod.POST)
     public String processEditForm(Model model, @ModelAttribute @Valid User user,
-                                  Errors errors, @RequestParam int userId,
-                                  @PathVariable int dateId, @ModelAttribute String food) {
+                                  Errors errors, @ModelAttribute Foodday date, @RequestParam int userId,
+                                  @PathVariable int dateId, @ModelAttribute String food,
+                                  @CookieValue(value="user", defaultValue = "none") String name) {
+
+        if(name.equals("none")) {
+            return "redirect:/user";
+        }
 
         if (errors.hasErrors()) {
 
@@ -187,9 +224,9 @@ public class FooddayController {
 
         User aUser = userDao.findOne(userId);
         aUser.setName(user.getName());
-        Foodday date = fooddayDao.findOne(dateId);
-        AddUserForm form = new AddUserForm(date, userDao.findAll(), food);
-        form.setFood(form.getFood());
+        Foodday aDate = fooddayDao.findOne(dateId);
+        //AddUserForm form = new AddUserForm(userDao.findAll(), food, date);
+        //aDate.setFood(aDate.getFood());
         userDao.save(aUser);
 
         return "redirect:/date/view/" + dateId;
